requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-background-tasks==1.2.5
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-colorfield==0.3.2
django-compat==1.0.15
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-unixtimestampfield==0.4.0
idna==2.10
install==1.3.3
Pillow==7.2.0
pusher-push-notifications==2.0.0
pyfcm==1.4.7
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
